fix: multiple exports issue
